refactor: extract FLOPs estimator from Megatron trainer to shared utility
